modularize protection flags into dedicated file
